UIL-95 feat: add AnimatedIcon component
