+++ packages/ui-kit/src/components/animated-icon/index.ts
@@ -1,0 +1,1 @@
+export * from './animated-icon'
