+++ packages/ui-kit/src/index.ts
@@ -9,6 +9,7 @@ export * from './utils'
 export type { IconProps } from './assets/icons/IconWrapper'
 
 export * from './components/alert'
+export * from './components/animated-icon'
 export * from './components/animated-info-icon'
 export * from './components/button'
 export * from './components/card'
