--- packages/ui-kit/stories/components/feedback/animated-info-icon/animated-info-icon.stories.tsx
@@ -1,18 +1,0 @@
-import { AnimatedInfoIcon } from '../../../../src'
-import { Meta, StoryObj } from '@storybook/react'
-
-const meta = {
-  component: AnimatedInfoIcon,
-  title: 'Components/Feedback/AnimatedInfoIcon',
-} satisfies Meta<typeof AnimatedInfoIcon>
-
-export default meta
-type Story = StoryObj<typeof AnimatedInfoIcon>
-
-export const Default: Story = {
-  args: { enableAnimation: true },
-}
-
-export const AnimationDisabled: Story = {
-  args: { enableAnimation: false },
-}
